LayoutTests/fast/dom/HTMLLinkElement/preconnect-support.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests that Link's rel=preconnect is reported as supported.");

const testLink = document.createElement("link");
testLink.rel = "preconnect";
testLink.href = "http://localhost:8000";

shouldBe("testLink.relList.length", "1");
shouldBeEqualToString("testLink.relList[0]", "preconnect");
shouldBeTrue("testLink.relList.supports('preconnect')");
</script>
</body>
</html>
