LayoutTests/fast/dom/HTMLLinkElement/preconnect-support-expected.txt

Tests that Link's rel=preconnect is reported as supported.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testLink.relList.length is 1
PASS testLink.relList[0] is "preconnect"
PASS testLink.relList.supports('preconnect') is true
PASS successfullyParsed is true

TEST COMPLETE
